+++ electron/src/settings/config.ts
